requirements.txt
newspaper3k
streamlit
numpy==1.24.3
pandas==1.5.3
scipy==1.10.1
spacy
en-core-web-sm
